backend-MindVarta/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai>=1.0.0
python-dotenv
gtts
psycopg2-binary
bcrypt
